% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-PlotGG-class.R
\name{addCustomPoints}
\alias{addCustomPoints}
\title{Add custom points (generic)}
\usage{
addCustomPoints(plot_obj, custom_points, ...)
}
\arguments{
\item{plot_obj}{A plot-like object.}

\item{custom_points}{A list/data for points.}

\item{...}{Passed on to methods.}
}
\description{
Add custom points (generic)
}
