% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{agreement_description}
\alias{agreement_description}
\title{adds the description of the RunTest_multiple_logistic_regression app}
\usage{
agreement_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_multiple_logistic_regression
}
\description{
adds the description of the RunTest_multiple_logistic_regression app
}
