% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{ANOVA_description}
\alias{ANOVA_description}
\title{adds the description of the RunTest_ANOVA_one_way app}
\usage{
ANOVA_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_ANOVA_one_way
}
\description{
adds the description of the RunTest_ANOVA_one_way app
}
