% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-applyFormat.R
\name{applyFormatUI}
\alias{applyFormatUI}
\alias{applyFormatServer}
\title{UI for applying format to selected elements}
\usage{
applyFormatUI(
  wrapper_id,
  format_FUN,
  label_selected = "Select point(s)",
  choices_selected = c(`Add a point ...` = ""),
  width = "100\%",
  ...
)

applyFormatServer(
  id,
  default_style,
  formatServerFUN,
  element_list = reactiveVal(),
  style_prefix = "",
  plot_type = c("ggplot", "base", "none"),
  layout_group = reactive(c()),
  group_entries = c(),
  ...
)
}
\arguments{
\item{wrapper_id}{character. Wrapper ID for the UI.}

\item{format_FUN}{function. Function to create the UI for the format.}

\item{label_selected}{character. Label for the selectInput.}

\item{choices_selected}{character. Choices for the selectInput.}

\item{width}{character. Width of the selectInput.}

\item{...}{further arguments passed to the formatServerFUN.}

\item{id}{character. Module ID.}

\item{default_style}{list. Default style settings.}

\item{formatServerFUN}{function. Function to create the server logic for the format.}

\item{element_list}{reactiveVal. List of elements to apply the format to.}

\item{style_prefix}{character. Prefix for the style settings.}

\item{plot_type}{character. Type of plot. Default is "ggplot".}

\item{layout_group}{reactive. Group of elements to apply the format to.}

\item{group_entries}{character. Specify entries for 'layout_group' to apply the format to.
If empty, all entries are updated for 'layout_group'.}
}
\description{
UI for applying format to selected elements

Server logic for applying format to selected elements
}
