% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{approxFormula}
\alias{approxFormula}
\title{the description of the "approx formula" field in the Marquardt - Levenberg application}
\usage{
approxFormula(language)
}
\arguments{
\item{accepts}{the language in which is written}
}
\value{
the description of the "approx formula" field in the selected language
}
\description{
the description of the "approx formula" field in the Marquardt - Levenberg application
}
