% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/areAreNot.R, R/saNieSa.R
\name{areAreNot}
\alias{areAreNot}
\title{function for returning a line to the report of ChiSquared tabel test and Fisher table test}
\usage{
areAreNot(P, language = "pl")

areAreNot(P, language = "pl")
}
\arguments{
\item{accepts}{a P value}

\item{accepts}{a P value}
}
\value{
returns a formatted string

returns a formatted string
}
\description{
function for returning a line to the report of ChiSquared tabel test and Fisher table test

function for returning a line to the report of ChiSquared tabel test and Fisher table test
}

