% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-PlotGG-class.R
\name{as.ggplot.PlotGG}
\alias{as.ggplot.PlotGG}
\title{Coerce PlotGG to ggplot}
\usage{
\method{as.ggplot}{PlotGG}(obj)
}
\arguments{
\item{obj}{A plot-like object.}
}
\value{
A `ggplot` object.
}
\description{
Coerce PlotGG to ggplot
}
