% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{BA_description}
\alias{BA_description}
\title{adds the description of the RunTest_Bland_Altman app}
\usage{
BA_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_Bland_Altman
}
\description{
adds the description of the RunTest_Bland_Altman app
}
