% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{backwardSelection}
\alias{backwardSelection}
\title{the description of the "multiple regression" backward selection field}
\usage{
backwardSelection(step, language)
}
\arguments{
\item{accepts}{the logical value for outputing "backward selection" and the language in which the app will be written}
}
\value{
the "backward selection" in the appropriate language or an empty string
}
\description{
the description of the "multiple regression" backward selection field
}

