% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{binom_trend_description}
\alias{binom_trend_description}
\title{adds the description of the RunTest_binom_trend}
\usage{
binom_trend_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_binom_trend
}
\description{
adds the description of the RunTest_binom_trend
}

