% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{boot_confint_description}
\alias{boot_confint_description}
\title{adds the description of the RunTest_boot_confint app}
\usage{
boot_confint_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_boot_confint
}
\description{
adds the description of the RunTest_boot_confint app
}
