% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-formatGGplot.R
\name{calculateRescalingFactors}
\alias{calculateRescalingFactors}
\title{Calculate Rescaling Factors}
\usage{
calculateRescalingFactors(oldLimits, newLimits = NULL)
}
\arguments{
\item{oldLimits}{(numeric) old limits}

\item{newLimits}{(numeric) new limits}
}
\value{
(list) with scale and center
}
\description{
Calculate Rescaling Factors
}
