% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{chi_squared_test_description}
\alias{chi_squared_test_description}
\title{adds the description of the RunTest_chi_squared}
\usage{
chi_squared_test_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_chi_squared
}
\description{
adds the description of the RunTest_chi_squared
}

