% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{colorList}
\alias{colorList}
\title{contains constantst used in building shiny apps, like list of colors, etc.}
\usage{
colorList(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
returns the color list in the appropriate language
}
\description{
contains constantst used in building shiny apps, like list of colors, etc.
}
