% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/convergesDoesNotConverge.R, R/zbiegaNieZbiega.R
\name{convergesDoesNotConverge}
\alias{convergesDoesNotConverge}
\title{reports whether the nls algorithm converges}
\usage{
convergesDoesNotConverge(modelNLS, language = "pl")

convergesDoesNotConverge(modelNLS, language = "pl")
}
\arguments{
\item{the}{nls model, the language}

\item{the}{nls model, the language}
}
\value{
either "converges" or "diverges" in the appropriate language

either "converges" or "diverges" in the appropriate language
}
\description{
reports whether the nls algorithm converges

reports whether the nls algorithm converges
}

