% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-convertToExpression.R
\name{convertToExpression}
\alias{convertToExpression}
\title{Convert to Expression}
\usage{
convertToExpression(string)
}
\arguments{
\item{string}{(character) string to convert}
}
\value{
expression
}
\description{
Convert a string to an expression
}
