% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{cor_matrix_p_description}
\alias{cor_matrix_p_description}
\title{adds the description of the RunTest_correlation_matrix_Pearson}
\usage{
cor_matrix_p_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_correlation_matrix_Pearson
}
\description{
adds the description of the RunTest_correlation_matrix_Pearson
}

