% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationMatrix.R
\name{corMatrix}
\alias{corMatrix}
\title{oblicza values p i cCoefolczynniki r/rho dla macierzy korelacji}
\usage{
corMatrix(myDataFrama, method = "p")
}
\arguments{
\item{myData}{jako dataframe}
}
\value{
macierz tekstowa z valuesami r/rho (gora) i p (dol)
}
\description{
oblicza values p i cCoefolczynniki r/rho dla macierzy korelacji
}

