% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/correlationMatrix.R
\name{correlationDraw}
\alias{correlationDraw}
\title{rysuje macierz korelacji wraz z linią loess - pozyczylem troche kodu z internetu - musze sprawdzic skad}
\usage{
correlationDraw(
  myDataFrama,
  tTitle = "",
  nNames = names(myDataFrama),
  method = "p"
)
}
\arguments{
\item{przyjmuje}{dataFrame z danymi, tTitle calej analizy, oraz nNames kolumn}
}
\value{
zwraca rysunek - correlation matrix z odpowiednio sformatowanymi panelami
}
\description{
rysuje macierz korelacji wraz z linią loess - pozyczylem troche kodu z internetu - musze sprawdzic skad
}
