% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{Cox_regression_description}
\alias{Cox_regression_description}
\title{adds the description of the RunTest_SurvivalSuite_Cox app}
\usage{
Cox_regression_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_SurvivalSuite_KM
}
\description{
adds the description of the RunTest_SurvivalSuite_Cox app
}

