% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/uiHelper.R
\name{creationInfo}
\alias{creationInfo}
\title{tekst informacyjny}
\usage{
creationInfo(
 
    tekst = "<br>Program napisany przez <a href=\\"http://imyData.pl\\" target=\\"_blank\\">Jarosław Piskorski-imyData.pl</a>."
)
}
\arguments{
\item{text}{tekst w HTML do wyswietlenia w panelu shiny'ego}
}
\value{
funkcja HTML
}
\description{
tekst informacyjny
}
