% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-customPoints.R
\name{customPointsUI}
\alias{customPointsUI}
\alias{customPointsServer}
\title{Custom Points Module}
\usage{
customPointsUI(id, title = "Custom Points", titleTag = "h4")

customPointsServer(id)
}
\arguments{
\item{id}{namespace id}

\item{title}{(character) module title}

\item{titleTag}{(character) HTML tag to put around the title, e.g. "h4" for \code{h4} from
\code{htmltools}}
}
\description{
This module provides the UI that allows to add, style and remove custom points, and returns the
 config list for the custom points.

Backend for custom points module
}
