% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dataExport.R
\name{dataExportButton}
\alias{dataExportButton}
\title{Data Export Button}
\usage{
dataExportButton(id, label = "Export Data")
}
\arguments{
\item{id}{module id}

\item{label}{(character) button label}
}
\description{
Data Export Button
}
