% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/decision.R, R/decyzja.R
\name{decision}
\alias{decision}
\title{takes the decision if a p value indicates statistical significance}
\usage{
decision(p, language = "pl", liczba = "singular")

decision(p, language = "pl", liczba = "singular")
}
\arguments{
\item{p}{-value, language}
}
\value{
string with the "is" "is not" decision

string with the "is" "is not" decision
}
\description{
takes the decision if a p value indicates statistical significance

takes the decision if a p value indicates statistical significance
}
