% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotTitles.R
\name{defaultTextFormat}
\alias{defaultTextFormat}
\title{Default Title Format}
\usage{
defaultTextFormat(type = c("none", "ggplot", "base"))
}
\description{
Initial values for title dependent on the plot type
}
