% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotTitles.R
\name{defaultTitleFormat}
\alias{defaultTitleFormat}
\title{Default Title Format}
\usage{
defaultTitleFormat(type = c("none", "ggplot", "base"))
}
\arguments{
\item{type}{(character) plot type, one of "ggplot" or "base"}
}
\description{
Initial values for title dependent on the plot type
}
