% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{dependentVariable}
\alias{dependentVariable}
\title{the description of the "dependent variable" field}
\usage{
dependentVariable(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the "dependent variable" field in the selected language
}
\description{
the description of the "dependent variable" field
}
