% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{describeMultiple}
\alias{describeMultiple}
\title{the description of the "multiple data entry" field}
\usage{
describeMultiple(language)
}
\arguments{
\item{accepts}{the language in which is written}
}
\value{
the description of the "independent variable" field in the selected language
}
\description{
the description of the "multiple data entry" field
}
