% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-dataExport.R
\name{exportXLSX}
\alias{exportXLSX}
\title{Export to xlsx}
\usage{
exportXLSX(file, dat)
}
\arguments{
\item{file}{filename}

\item{dat}{(list) If \code{dat} is a data.frame \code{dat} will be written into a single sheet
"Sheet 1". If \code{dat} is a named list of data.frames \code{dat} each data.frame is written
 into a separate sheet with sheet names taken from the names of the list.}
}
\description{
Export a single data.frame, or export a list of data.frames into workbook containing one
data.frame in each sheet of a workbook
}
