% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-shinyTryCatch.R
\name{extractSingleTryCatch}
\alias{extractSingleTryCatch}
\title{Extract Single Try Catch}
\usage{
extractSingleTryCatch(condition, type)
}
\arguments{
\item{condition}{error or warning object}

\item{type}{(character) type of the message: "error" or "warning"}
}
\description{
Extract Single Try Catch
}
