% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotTitles.R
\name{fontChoices}
\alias{fontChoices}
\title{Font Choices}
\usage{
fontChoices(type = c("none", "ggplot", "base"))
}
\arguments{
\item{type}{(character) plot type, one of "ggplot" or "base"}
}
\description{
Font Choices
}
