% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-formatText.R
\name{fontChoicesSelect}
\alias{fontChoicesSelect}
\title{Font Choices}
\usage{
fontChoicesSelect(type = c("ggplot", "base"))
}
\arguments{
\item{type}{(character) plot type, one of "ggplot" or "base"}
}
\description{
Mapping of font choices dependent on the plot type
}
