% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-formatGGplot.R
\name{formatAxisOfGGplot}
\alias{formatAxisOfGGplot}
\title{Format an axis Of a ggplot}
\usage{
formatAxisOfGGplot(plot, axisFormat, axis = c("x", "y"), ...)
}
\arguments{
\item{plot}{(ggplot)}

\item{axisFormat}{(list) named list with axis format definitions for a specific axis}

\item{axis}{(character) axis to format, one of "xAxis", "yAxis"}

\item{...}{additional arguments for \code{scale_x_continuous} or \code{scale_y_continuous}}
}
\description{
Format an axis Of a ggplot
}
