% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-PlotGG-class.R
\name{formatLegend.PlotGG}
\alias{formatLegend.PlotGG}
\title{Format legend for PlotGG}
\usage{
\method{formatLegend}{PlotGG}(plot_obj, legend, scaleFUN = ggplot2::scale_color_manual, ...)
}
\arguments{
\item{plot_obj}{(ggplot)}

\item{legend}{(list) named list with style definitions, or output of \code{plotLegendServer}}

\item{scaleFUN}{(function) function to set scale, e.g. \code{ggplot2::scale_color_manual}}

\item{...}{additional element specifications not part of base ggplot2. In general,
these should also be defined in the \verb{element tree} argument. \link[rlang:splice]{Splicing} a list is also supported.}
}
\value{
A `PlotGG` object (modified).
}
\description{
Format legend for PlotGG
}
