% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-PlotGG-class.R
\name{formatLegend}
\alias{formatLegend}
\title{Format legend (generic)}
\usage{
formatLegend(plot_obj, legend, ...)
}
\arguments{
\item{plot_obj}{A plot-like object.}

\item{legend}{A legend configuration list.}

\item{...}{Passed on to methods.}
}
\description{
Format legend (generic)
}
