% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-formatLine.R
\name{formatLineUI}
\alias{formatLineUI}
\alias{formatLineServer}
\title{Format Line UI}
\usage{
formatLineUI(
  id,
  title = NULL,
  titleTag = "h4",
  initStyle = defaultLineFormat()
)

formatLineServer(
  id,
  hideInput = c(),
  initStyle = defaultLineFormat(),
  reloadInit = reactiveVal(FALSE)
)
}
\arguments{
\item{id}{namespace id}

\item{initStyle}{(list) optional, named list with style definitions, should have the same format
as the default output of \code{plotPointsServer}}

\item{hideInput}{(character) inputs that should be disabled (hidden) when applying this module.
Possible inputs are "hide", "symbol", "color", "size", "alpha", "colorBg", "lineWidthBg".
Please use \code{shinyjs::useShinyjs()} in your UI function to enable this feature.}

\item{reloadInit}{(reactiveVal) logical, should \code{initStyle} be reloaded?}
}
\description{
Backend for format line module
}
