% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-formatGGplot.R
\name{formatPointLabelsOfGGPlot}
\alias{formatPointLabelsOfGGPlot}
\title{Format Point Labels Of GGplot}
\usage{
formatPointLabelsOfGGPlot(
  plot,
  data,
  labelStyle = getLabelStyle("ggplot"),
  ...
)
}
\arguments{
\item{plot}{(ggplot)}

\item{data}{(data.frame) data with x, y and label information}

\item{labelStyle}{(list) named list with style definitions}

\item{...}{(list) arguments for \code{geom_text}}
}
\description{
Format Point Labels Of GGplot
}
