% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-formatGGplot.R
\name{formatRangesOfGGplot}
\alias{formatRangesOfGGplot}
\title{Axes Ranges Of GGplot (deprecated)}
\usage{
formatRangesOfGGplot(...)
}
\arguments{
\item{...}{arguments of \code{formatScalesOfGGplot}}
}
\description{
Axes Ranges Of GGplot (deprecated)
}
