% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-PlotGG-class.R
\name{formatScales.PlotGG}
\alias{formatScales.PlotGG}
\title{Format scales for PlotGG}
\usage{
\method{formatScales}{PlotGG}(
  plot_obj,
  ranges,
  xLabels = NULL,
  yLabels = NULL,
  ySecAxisTitle = NULL,
  ...
)
}
\arguments{
\item{plot_obj}{(ggplot)}

\item{ranges}{(list) named list with range definitions, output of \code{plotRangesServer}}

\item{xLabels}{(list) named list with x axis labels, e.g. \code{list(breaks = c(1, 2, 3), labels = c("A", "B", "C"))}}

\item{yLabels}{(list) named list with y axis labels, e.g. \code{list(breaks = c(1, 2, 3), labels = c("A", "B", "C"))}}

\item{ySecAxisTitle}{(character) title of secondary y axis}

\item{...}{Passed on to [formatScales.PlotGG()].}
}
\value{
A `PlotGG` object (modified).
}
\description{
Format scales for PlotGG
}
