% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-PlotGG-class.R
\name{formatScales}
\alias{formatScales}
\title{Format plot scales (generic)}
\usage{
formatScales(plot_obj, ranges, ...)
}
\arguments{
\item{plot_obj}{A plot-like object.}

\item{ranges}{A named list with axis ranges/transforms.}

\item{...}{Passed on to methods.}
}
\description{
Format plot scales (generic)
}
