% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-PlotGG-class.R
\name{formatTitles.PlotGG}
\alias{formatTitles.PlotGG}
\title{Format titles for PlotGG}
\usage{
\method{formatTitles}{PlotGG}(plot_obj, text, ...)
}
\arguments{
\item{plot_obj}{(ggplot)}

\item{text}{(list) named list with title definitions, output of \code{plotTitlesServer}}
}
\value{
A `PlotGG` object (modified).
}
\description{
Format titles for PlotGG
}
