% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-PlotGG-class.R
\name{formatTitles}
\alias{formatTitles}
\title{Format plot titles (generic)}
\usage{
formatTitles(plot_obj, text, ...)
}
\arguments{
\item{plot_obj}{A plot-like object.}

\item{text}{A named list with title/text specs.}

\item{...}{Passed on to methods.}
}
\description{
Format plot titles (generic)
}
