% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-formatGGplot.R
\name{formatTitlesOfGGplot}
\alias{formatTitlesOfGGplot}
\title{Format Titles Of GGplot}
\usage{
formatTitlesOfGGplot(plot, titles)
}
\arguments{
\item{plot}{(ggplot)}

\item{titles}{(list) named list with title definitions, output of \code{plotTitlesServer}}
}
\description{
Format Titles Of GGplot
}
