% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConfInt.R
\name{getConfInt}
\alias{getConfInt}
\title{takes a 1X2 vector and formats it as a string with confidence interval}
\usage{
getConfInt(x)
}
\arguments{
\item{x}{-1x2 vector}
}
\value{
string confidence interval in brackets
}
\description{
takes a 1X2 vector and formats it as a string with confidence interval
}

