% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getConfInt.R
\name{getConfInt2}
\alias{getConfInt2}
\title{takes a nX2 vector and formats it as a string vector with confidence interval}
\usage{
getConfInt2(x, y)
}
\arguments{
\item{x}{-nx2 vector}
}
\value{
string confidence interval in brackets vector
}
\description{
takes a nX2 vector and formats it as a string vector with confidence interval
}

