% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-formatGGplot.R
\name{getLabelStyle}
\alias{getLabelStyle}
\title{Get Default Point Style}
\usage{
getLabelStyle(type = c("ggplot", "base"))
}
\arguments{
\item{type}{(character) type of plot to add labels to, either 'ggplot' or 'base'}
}
\description{
Get Default Point Style
}
