% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPFromMultiple.R
\name{getPFromMultiple}
\alias{getPFromMultiple}
\title{takes a multiple linear regression model and calculates the p-value for the whole model}
\usage{
getPFromMultiple(model)
}
\arguments{
\item{lm}{model}
}
\value{
numerical p-value
}
\description{
takes a multiple linear regression model and calculates the p-value for the whole model
}
