% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getPFromMultipleGLM.R
\name{getPFromMultipleGLM}
\alias{getPFromMultipleGLM}
\title{takes a multiple logistic regression model and returns the string stating whether the model contains any statistically significant covariates - as a rule there are no whole-model tests or p-values for generalized linear models}
\usage{
getPFromMultipleGLM(model, lang = "pl")
}
\arguments{
\item{glm}{model}
}
\value{
string stating whether the model contains any statistically significant covariates
}
\description{
takes a multiple logistic regression model and returns the string stating whether the model contains any statistically significant covariates - as a rule there are no whole-model tests or p-values for generalized linear models
}

