% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-headerButtons.R
\name{headerButtonsUI}
\alias{headerButtonsUI}
\title{Add header buttons}
\usage{
headerButtonsUI(id, help_link)
}
\arguments{
\item{id}{module id}

\item{help_link}{link that is opened when help button is clicked}
}
\value{
tagList with divs
}
\description{
Add header buttons
}
