% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{independentVariable}
\alias{independentVariable}
\title{the description of the "independent variable" field}
\usage{
independentVariable(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the description of the "independent variable" field in the selected language
}
\description{
the description of the "independent variable" field
}

