% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/isIsNot.R, R/jestNieJest.R
\name{isIsNot}
\alias{isIsNot}
\title{function for returning a line to the report of ChiSquared tabel test and Fisher table test}
\usage{
isIsNot(P, language = "pl")

isIsNot(P, language = "pl")
}
\arguments{
\item{accepts}{a P value}
}
\value{
returns a formatted string

returns a formatted string
}
\description{
function for returning a line to the report of ChiSquared tabel test and Fisher table test

function for returning a line to the report of ChiSquared tabel test and Fisher table test
}
