% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotTitles.R
\name{keep_deepest_names}
\alias{keep_deepest_names}
\title{Keep Deepest Names}
\usage{
keep_deepest_names(x)
}
\arguments{
\item{x}{(vector)  A named vector with named elements and nested names separated by '.'}
}
\value{
A character vector containing the names of the deepest level elements.
}
\description{
Extracts the names of the deepest level elements from nested names.
}
