% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{kolor}
\alias{kolor}
\title{the description of the "color select" field}
\usage{
kolor(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the "color select" field in the selected language
}
\description{
the description of the "color select" field
}

