% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{Kruskall_Wallis_description}
\alias{Kruskall_Wallis_description}
\title{adds the description of the RunTest_Kruskall_Wallis_test app}
\usage{
Kruskall_Wallis_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_Kruskall_Wallis_test
}
\description{
adds the description of the RunTest_Kruskall_Wallis_test app
}

