% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{loadFile}
\alias{loadFile}
\title{the description of the "file upload" field}
\usage{
loadFile(language, typ)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the "file upload name ad type" field in the selected language
}
\description{
the description of the "file upload" field
}
