% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{logistic_regression_2x2_description}
\alias{logistic_regression_2x2_description}
\title{adds the description of the RunTest_log_contingency app}
\usage{
logistic_regression_2x2_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_log_contingency
}
\description{
adds the description of the RunTest_log_contingency app
}

