% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{multiple_prop_test_description}
\alias{multiple_prop_test_description}
\title{adds the description of the RunTest_prop_many app}
\usage{
multiple_prop_test_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_prop_many
}
\description{
adds the description of the RunTest_prop_many app
}
