% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{multiple_regression_description}
\alias{multiple_regression_description}
\title{adds the description of the RunTest_multiple_regression app}
\usage{
multiple_regression_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_multiple_regression
}
\description{
adds the description of the RunTest_multiple_regression app
}

