% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-PlotGG-class.R
\name{new_PlotGG}
\alias{new_PlotGG}
\title{GGPlot Formatting Class}
\usage{
new_PlotGG(plot)
}
\arguments{
\item{plot}{A ggplot object.}
}
\value{
An object of class `PlotGG`.
}
\description{
This script defines the S3 class `PlotGG` and its methods for formatting ggplot2 plots.
This class is a wrapper around the functions for formatting ggplots stored in 03-formatGGplot.R
Constructor for PlotGG
}
\details{
This function creates a new instance of the PlotGG class.
}
