% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{nonlinear_regression_description}
\alias{nonlinear_regression_description}
\title{adds the description of the RunTest_nonlinear_reg app}
\usage{
nonlinear_regression_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_nonlinear_reg
}
\description{
adds the description of the RunTest_nonlinear_reg app
}

