% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{normalitytest_description}
\alias{normalitytest_description}
\title{adds the description of the RunTest_normality_test app}
\usage{
normalitytest_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_normality_test
}
\description{
adds the description of the RunTest_normality_test app
}

