% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotPoints.R
\name{observeAndUpdatePointElements}
\alias{observeAndUpdatePointElements}
\title{Observe Point Elements}
\usage{
observeAndUpdatePointElements(input, output, session, style, type)
}
\arguments{
\item{input}{input object from server function}

\item{output}{output object from server function}

\item{session}{session from server function}

\item{style}{(reactiveValue) contains point elements}

\item{type}{(character) Type of the plot to edit points for, one of "ggplot", "base".}
}
\description{
Observe inputs for different text elements (e.g. color, size, ...) of a selected label (e.g.
plot title, axis text, ...) and store values in the reactiveValues list 'text'.
}
