% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotRanges.R
\name{observeAndUpdateRangeElementsOfLabel}
\alias{observeAndUpdateRangeElementsOfLabel}
\title{Observe Range Elements Of Label}
\usage{
observeAndUpdateRangeElementsOfLabel(input, output, session, ranges)
}
\arguments{
\item{input}{input object from server function}

\item{output}{output object from server function}

\item{session}{session from server function}

\item{ranges}{(reactiveValue) contains range elements}
}
\description{
Observe inputs for different text elements (e.g. color, size, ...) of a selected label (e.g.
plot title, axis text, ...) and store values in the reactiveValues list 'text'.
}
