% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotTitles.R
\name{observeAndUpdateTextElementsOfLabel}
\alias{observeAndUpdateTextElementsOfLabel}
\title{Observe Text Elements Of Label}
\usage{
observeAndUpdateTextElementsOfLabel(input, output, session, text)
}
\arguments{
\item{input}{input object from server function}

\item{output}{output object from server function}

\item{session}{session from server function}

\item{text}{(reactiveValue) contains text elements}
}
\description{
Observe inputs for different text elements (e.g. color, size, ...) of a selected label (e.g.
plot title, axis text, ...) and store values in the reactiveValues list 'text'.
}
