% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{odds_ratio_test_description}
\alias{odds_ratio_test_description}
\title{adds the description of the RunTest_Fisher}
\usage{
odds_ratio_test_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_Fisher
}
\description{
adds the description of the RunTest_Fisher
}
