% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{onesummary_description}
\alias{onesummary_description}
\title{adds the description of the RunTest_onesummary app}
\usage{
onesummary_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_normality_test
}
\description{
adds the description of the RunTest_onesummary app
}
