% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{ordered_logistic_regression_description}
\alias{ordered_logistic_regression_description}
\title{adds the description of the RunTest_ordinal_logistic_regression app}
\usage{
ordered_logistic_regression_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_ordinal_logistic_regression
}
\description{
adds the description of the RunTest_ordinal_logistic_regression app
}

