% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{Pearson_correlation_description}
\alias{Pearson_correlation_description}
\title{adds the description of the RunTest_Pearson_correlation app}
\usage{
Pearson_correlation_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_Pearson_correlation
}
\description{
adds the description of the RunTest_Pearson_correlation app
}

