% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotExport.R
\name{plotExportButton}
\alias{plotExportButton}
\title{Plot Export Button}
\usage{
plotExportButton(id, label = "Export Plot")
}
\arguments{
\item{id}{module id}

\item{label}{(character) button label}
}
\value{
actionButton
}
\description{
Plot Export Button
}
