% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotLegend.R
\name{plotLegendUI}
\alias{plotLegendUI}
\alias{plotLegendServer}
\title{Plot Legend UI}
\usage{
plotLegendUI(id, title = NULL, titleTag = "h4")

plotLegendServer(id)
}
\arguments{
\item{id}{namespace id}

\item{title}{(character) title of the module}

\item{titleTag}{(character) HTML tag to put around the title, e.g. "h4" for \code{h4} from
\code{htmltools}}
}
\description{
Function to create the UI for the legend

Function to create the server for the legend
}
