% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotPoints.R
\name{plotPointsUI}
\alias{plotPointsUI}
\title{Plot Points UI}
\usage{
plotPointsUI(
  id,
  title = "Data Points",
  type = c("ggplot", "base"),
  initStyle = NULL
)
}
\arguments{
\item{id}{module id}

\item{title}{(character) module title}

\item{type}{(character) Type of the plot to edit points for, one of "ggplot", "base".}

\item{initStyle}{(list) optional, named list with style definitions, should have the same format
as the default output of \code{plotPointsServer}}
}
\value{
tagList
}
\description{
Plot Points UI
}
