% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotRanges.R
\name{plotRangesServer}
\alias{plotRangesServer}
\title{Server function for plot ranges}
\usage{
plotRangesServer(
  id,
  type = c("none", "ggplot", "base"),
  initRanges = NULL,
  userRanges = NULL,
  axes = c(`x axis` = "xAxis", `y axis` = "yAxis")
)
}
\arguments{
\item{id}{namespace id}

\item{type}{(character) type of the plot to add ranges to, one of "ggplot", "base".}

\item{initRanges}{(list) optional, named list with range definitions}

\item{userRanges}{(reactiveValues) optional, reactiveValues object containing ranges for the axes.}

\item{axes}{(character) named vector of axes to add ranges to, e.g.
\code{c("x axis" = "xAxis", "y axis" = "yAxis", "2nd y axis" = "yAxis2")}.}
}
\description{
Backend for plot ranges module
}
