% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-formatText.R, R/02-plotTitles.R
\name{formatTextUI}
\alias{formatTextUI}
\alias{plotTitlesUI}
\alias{plotTitlesServer}
\title{Text Format UI}
\usage{
formatTextUI(
  id,
  type = c("ggplot", "base", "none"),
  initTitle = NULL,
  initAxis = NULL
)

plotTitlesUI(
  id,
  title = "Plot Texts",
  titleTag = "h4",
  type = c("ggplot", "base", "none"),
  initText = NULL
)

plotTitlesServer(
  id,
  type = c("none", "ggplot", "base"),
  availableElements = c("title", "axis"),
  showParseButton = TRUE,
  initText = NULL
)
}
\arguments{
\item{id}{namespace id}

\item{type}{(character) Type of the plot to add titles to, one of "none", "ggplot", "base".}

\item{title}{(character) module title}

\item{titleTag}{(character) HTML tag to put around the title, e.g. "h4" for \code{h4} from
\code{htmltools}}

\item{initText}{(list) optional, named list with title definitions, or output of \code{plotTitlesServer}}

\item{availableElements}{(character) set of available labels for specifying the format of text.
May contain elements from \code{c("title", "axis", "yaxis2", "legend")}.}

\item{showParseButton}{(logical) Show parse button for parsing mathematical expressions.}
}
\value{
tagList

tagList
}
\description{
Backend for plot titles module
}
