% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotTitles.R
\name{plotTitlesUI}
\alias{plotTitlesUI}
\title{Plot Titles UI}
\usage{
plotTitlesUI(
  id,
  title = "Plot Texts",
  type = c("ggplot", "base"),
  initText = NULL
)
}
\arguments{
\item{id}{module id}

\item{title}{(character) module title}

\item{type}{(character) Type of the plot to add titles to, one of "none", "ggplot", "base".}

\item{initText}{(list) optional, named list with title definitions, or output of \code{plotTitlesServer}}
}
\value{
tagList
}
\description{
Plot Titles UI
}
