% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/polyNew.R
\name{polyNew}
\alias{polyNew}
\title{this is just the poly function without the safeguards on the degree of polyomial, which some consider a bug - in this way I am able to fit the polynomial of order n-1 (n - number of datapoints)}
\usage{
polyNew(x, ..., degree = 1, coefs = NULL, raw = FALSE)
}
\arguments{
\item{x-vector}{and polynomial degree}
}
\value{
whatever poly returns
}
\description{
this is just the poly function without the safeguards on the degree of polyomial, which some consider a bug - in this way I am able to fit the polynomial of order n-1 (n - number of datapoints)
}
