% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{polynomial_regression_description}
\alias{polynomial_regression_description}
\title{adds the description of the RunTest_polynomial_regression app}
\usage{
polynomial_regression_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_polynomial_regression
}
\description{
adds the description of the RunTest_polynomial_regression app
}
