% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/puj.R
\name{puj}
\alias{puj}
\title{function for formating the p-value as it is inserted into text documents}
\usage{
puj(p, rmarkdown = TRUE)
}
\arguments{
\item{accepts}{the numerical p-value and the boolean value indicating whether the output will be used in a RMarkdown document}
}
\value{
returns a formated string corresponding to the p-value and RMarkdown setup
}
\description{
function for formating the p-value as it is inserted into text documents
}

