% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCategoricalArray.R
\name{readCategoricalArray}
\alias{readCategoricalArray}
\title{function for reading a data frame copied and pasted from Excel to numericInput shiny field}
\usage{
readCategoricalArray(numericInputString)
}
\arguments{
\item{accepts}{a string which is the result of pasting a selection from Excel}
}
\value{
returns a data frame corresponding to the selection in Excel. All entriest are strings
}
\description{
function for reading a data frame copied and pasted from Excel to numericInput shiny field
}

