% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readCategoricalVector.R
\name{readCategoricalVector}
\alias{readCategoricalVector}
\title{function for reading a vector copied and pasted from Excel to numericInput shiny field}
\usage{
readCategoricalVector(numericInputString)
}
\arguments{
\item{accepts}{a string which is the result of pasting a selection from Excel}
}
\value{
returns a data frame corresponding to the selection in Excel. All entriest are strings
}
\description{
function for reading a vector copied and pasted from Excel to numericInput shiny field
}
