% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odczytaj.z.okienka.R,
%   R/odczytajWierszZOkienka.R, R/readFromField.R, R/readLineFromField.R
\name{readFactorsFromField}
\alias{readFactorsFromField}
\title{odczytuje faktory wklejone lub wpisane do okienka}
\usage{
readFactorsFromField(listArgument)

readFactorsFromField(listArgument)

readFactorsFromField(listArgument)

readFactorsFromField(listArgument)
}
\arguments{
\item{listArgument}{zawiera liste zwracana przez shiny'ego, zawierajaca to co uzytkownik wkleil/wpisal do okienka}
}
\value{
zwraca wektor zawierajacy faktory z okienka

zwraca wektor zawierajacy faktory z okienka

zwraca wektor zawierajacy faktory z okienka

zwraca wektor zawierajacy faktory z okienka
}
\description{
odczytuje faktory wklejone lub wpisane do okienka

odczytuje faktory wklejone lub wpisane do okienka

odczytuje faktory wklejone lub wpisane do okienka

odczytuje faktory wklejone lub wpisane do okienka
}
