% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odczytajMacierz.R, R/readMatrix.R
\name{readMatrix}
\alias{readMatrix}
\title{function for reading a data frame copied and pasted from Excel to numericInput shiny field}
\usage{
readMatrix(numericInputString)

readMatrix(numericInputString)
}
\arguments{
\item{accepts}{a string which is the result of pasting a selection from Excel}

\item{accepts}{a string which is the result of pasting a selection from Excel}
}
\value{
returns a data frame corresponding to the selection in Excel. Interprets whther a column is numeric or factor

returns a data frame corresponding to the selection in Excel. Interprets whther a column is numeric or factor
}
\description{
function for reading a data frame copied and pasted from Excel to numericInput shiny field

function for reading a data frame copied and pasted from Excel to numericInput shiny field
}

