% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odczytaj.z.okienka.R, R/readFromField.R
\name{readNumbersFromField}
\alias{readNumbersFromField}
\title{odczytuje numbers wklejone lub wpisane do okienka, numbers moga byc z przecinkiem lub kropka}
\usage{
readNumbersFromField(listArgument)

readNumbersFromField(listArgument)
}
\arguments{
\item{listArgument}{zawiera liste zwracana przez shiny'ego, zawierajaca to co uzytkownik wkleil/wpisal do okienka}

\item{listArgument}{zawiera liste zwracana przez shiny'ego, zawierajaca to co uzytkownik wkleil/wpisal do okienka}
}
\value{
zwraca wektor zawierajacy numbers z okienka

zwraca wektor zawierajacy numbers z okienka
}
\description{
odczytuje numbers wklejone lub wpisane do okienka, numbers moga byc z przecinkiem lub kropka

odczytuje numbers wklejone lub wpisane do okienka, numbers moga byc z przecinkiem lub kropka
}

