% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odczytajTabelke.R, R/readTable.R
\name{readTable}
\alias{readTable}
\title{function for reading a contingency table copied and pasted from Excel to numericInput shiny field}
\usage{
readTable(numericInputString)

readTable(numericInputString)
}
\arguments{
\item{accepts}{a string which is the result of pasting a selection from Excel}
}
\value{
returns a contingency table (matrix) corresponding to the selection in Excel .

returns a contingency table (matrix) corresponding to the selection in Excel .
}
\description{
function for reading a contingency table copied and pasted from Excel to numericInput shiny field

function for reading a contingency table copied and pasted from Excel to numericInput shiny field
}
