% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/odczytajTabelke2by2.R, R/readTable2by2.R
\name{readTable2by2}
\alias{readTable2by2}
\title{function for reading a 2x2 table with row and column names}
\usage{
readTable2by2(numericInputString)

readTable2by2(numericInputString)
}
\arguments{
\item{accepts}{a string which is the result of pasting a selection from Excel}
}
\value{
returns a matrix corresponding to the selection in Excel - a

returns a matrix corresponding to the selection in Excel - a
}
\description{
function for reading a 2x2 table with row and column names

function for reading a 2x2 table with row and column names
}
