% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-resolveValue.R
\name{resolveValue}
\alias{resolveValue}
\title{Resolve Value}
\usage{
resolveValue(var)
}
\arguments{
\item{var}{The variable to resolve, can be reactive or not}
}
\value{
The resolved value
}
\description{
Function to resolve the value, whether reactive or not
}
