% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{ROC_AUC_description}
\alias{ROC_AUC_description}
\title{adds the description of the RunTest_AUC app}
\usage{
ROC_AUC_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_AUC
}
\description{
adds the description of the RunTest_AUC app
}

