% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{sampleQuantilesDescription}
\alias{sampleQuantilesDescription}
\title{the description of the "sample quantiles" axis}
\usage{
sampleQuantilesDescription(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the "sample quantiles" axis in the selected language
}
\description{
the description of the "sample quantiles" axis
}

