% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{sensit_description}
\alias{sensit_description}
\title{adds the description of the RunTest_sensit}
\usage{
sensit_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_sensit
}
\description{
adds the description of the RunTest_sensit
}

