% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-formatGGplot.R
\name{setCustomTitle}
\alias{setCustomTitle}
\title{Set Custom Title}
\usage{
setCustomTitle(plot, labFun, ...)
}
\arguments{
\item{plot}{(ggplot)}

\item{labFun}{(function) function to set label, e.g. \code{xlab}}

\item{...}{(list) arguments for \code{labFun}}
}
\description{
Set label of a ggplot object
}
