% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-formatGGplot.R
\name{setLegendThemeOfGGplot}
\alias{setLegendThemeOfGGplot}
\title{Legend Theme Of GGplot}
\usage{
setLegendThemeOfGGplot(plot, legend, ...)
}
\arguments{
\item{plot}{(ggplot)}

\item{legend}{(list) named list with style definitions, or output of \code{plotLegendServer}}

\item{...}{(list) arguments for \code{theme}}
}
\description{
Apply theme settings for legend to a ggplot object.
}
