% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-shinyTryCatch.R
\name{shinyTryCatch}
\alias{shinyTryCatch}
\title{Shiny Try Catch}
\usage{
shinyTryCatch(
  expr,
  errorTitle = "Modeling failed",
  warningTitle = "",
  alertStyle = "shinyjs",
  inShiny = TRUE,
  suppressWarnings = FALSE
)
}
\arguments{
\item{expr}{expression to be evaluated.}

\item{errorTitle}{(character) error message title.}

\item{warningTitle}{(character) warning message title.}

\item{alertStyle}{(character) Either "shinyjs", or "shinyalert". Specifies how an error or a
warning is given out. If "shinyjs" than shinyjs::alert is used; if "shinyalert" than
shinyalert::shinyalert is used.}

\item{inShiny}{(logical) if FALSE prevents an alert in shiny, but instead a warning will be
displayed. Use this when not applied within a Shiny session.}

\item{suppressWarnings}{(logical) if TRUE, warnings are suppressed and not given out as an alert.}
}
\description{
Catch multiple errors and warnings, and forward them as an alert. If an error occurs, NULL is
returned. If a warning occurs, the result is returned.
Please, add \code{shinyjs::useShinyjs()} to the UI to enable the alerts.
}
