% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{single_categorical_description}
\alias{single_categorical_description}
\title{adds the description of the RunTest_summary_discrete_single app}
\usage{
single_categorical_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_summary_discrete_single
}
\description{
adds the description of the RunTest_summary_discrete_single app
}

