% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotTitles.R
\name{sizeValues}
\alias{sizeValues}
\title{Size Values}
\usage{
sizeValues(type = c("ggplot", "base"))
}
\arguments{
\item{type}{(character) plot type, one of "ggplot" or "base"}
}
\description{
Initial values for sliderInput title 'size' dependent on the plot type
}
