% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotTitles.R
\name{sizeValuesSlider}
\alias{sizeValuesSlider}
\title{Size Values Slider}
\usage{
sizeValuesSlider(type = c("ggplot", "base"))
}
\arguments{
\item{type}{(character) plot type, one of "ggplot" or "base"}
}
\description{
Initial values for sliderInput title 'size' dependent on the plot type
}
