% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maleWartosci.R, R/smallValues.R
\name{smallValues}
\alias{smallValues}
\title{function for returning a line to the report of ChiSquared if there are small values in the analysed table}
\usage{
smallValues(small, language = "pl")

smallValues(small, language = "pl")
}
\arguments{
\item{accepts}{a logical variable, TRUE means that there are small (<5) values in the table}

\item{accepts}{a logical variable, TRUE means that there are small (<5) values in the table}
}
\value{
returns a formatted string

returns a formatted string
}
\description{
function for returning a line to the report of ChiSquared if there are small values in the analysed table

function for returning a line to the report of ChiSquared if there are small values in the analysed table
}

