% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{Spearman_correlation_description}
\alias{Spearman_correlation_description}
\title{adds the description of the RunTest_Spearman_correlation app}
\usage{
Spearman_correlation_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_Spearman_correlation
}
\description{
adds the description of the RunTest_Spearman_correlation app
}

