% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{Survival_KM_description}
\alias{Survival_KM_description}
\title{adds the description of the RunTest_SurvivalSuite_KM app}
\usage{
Survival_KM_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_SurvivalSuite_KM
}
\description{
adds the description of the RunTest_SurvivalSuite_KM app
}

