% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotPoints.R
\name{symbolChoicesSelect}
\alias{symbolChoicesSelect}
\title{Font Choices}
\usage{
symbolChoicesSelect()
}
\description{
Mapping of font choices to numeric values used in \code{ggplot2::geom_point} or in \code{base::points()}.
}
