% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/multivariateTableCox.R
\name{tableCoxSummary}
\alias{tableCoxSummary}
\title{takes the summary of a Cox model and formats a table with results as wel as plotting the forest plot}
\usage{
tableCoxSummary(result, shouldIPlot = "no")
}
\arguments{
\item{tableCoxSummary}{accepts summary(coxModel)}
}
\value{
table, plot
}
\description{
takes the summary of a Cox model and formats a table with results as wel as plotting the forest plot
}
