% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/GEEmodelTable.R
\name{tableGEESummary}
\alias{tableGEESummary}
\title{takes the summary of a gee (generalized estimating equations) model and formats a table}
\usage{
tableGEESummary(result_summary, round_to = 3)
}
\arguments{
\item{geeSummary}{accepts summary(gee_model)}
}
\value{
table
}
\description{
takes the summary of a gee (generalized estimating equations) model and formats a table
}

