% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{theoreticalQuantilesDescription}
\alias{theoreticalQuantilesDescription}
\title{the description of the "theoretical quantiles" axis}
\usage{
theoreticalQuantilesDescription(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the "theoretical quantiles" axis in the selected language
}
\description{
the description of the "theoretical quantiles" axis
}

