% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{three_way_description}
\alias{three_way_description}
\title{adds the description of the RunTest_three_way app}
\usage{
three_way_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_three_way
}
\description{
adds the description of the RunTest_three_way app
}
