% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{ttest_single_description}
\alias{ttest_single_description}
\title{adds the description of the RunTest_t_test_single app}
\usage{
ttest_single_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_t_test_single
}
\description{
adds the description of the RunTest_t_test_single app
}

