% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-updateUserInputs.R
\name{updateUserInputs}
\alias{updateUserInputs}
\title{Update User Inputs}
\usage{
updateUserInputs(input, output, session, userInputs)
}
\arguments{
\item{input}{input object from server function}

\item{output}{output object from server function}

\item{session}{session from server function}

\item{userInputs}{(list) named list of inputs to be updated}
}
\description{
Update values from 'input' with all values from userInputs
}
