% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{variableI}
\alias{variableI}
\title{the description of the "variable no i" field}
\usage{
variableI(language, i)
}
\arguments{
\item{accepts}{the language in which the app will be written and the number of the variable}
}
\value{
the description of the "variable no i" field in the selected language
}
\description{
the description of the "variable no i" field
}
