% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{variableName}
\alias{variableName}
\title{the description of the "variable name" field}
\usage{
variableName(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the "variable name" field in the selected language
}
\description{
the description of the "variable name" field
}
