% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-vectorInput.R
\name{vectorInputServer}
\alias{vectorInputServer}
\title{Backend for Vector Input module}
\usage{
vectorInputServer(id, defaultInputs)
}
\arguments{
\item{id}{(character) The namespace identifier for the module}

\item{defaultInputs}{(reactive) Numeric vector containing the default values for the vector}
}
\description{
Server function for handling the logic of the numeric vector input module
}
