% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{wtest_single_description}
\alias{wtest_single_description}
\title{adds the description of the RunTest_wilcoxon_test_single app}
\usage{
wtest_single_description(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the program RunTest_wilcoxon_test
}
\description{
adds the description of the RunTest_wilcoxon_test_single app
}

