% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{xDescription}
\alias{xDescription}
\title{the description of the "x name" field}
\usage{
xDescription(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the "x name" field in the selected language
}
\description{
the description of the "x name" field
}

