% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/constants.R
\name{yDescription}
\alias{yDescription}
\title{the description of the "y name" field}
\usage{
yDescription(language)
}
\arguments{
\item{accepts}{the language in which the app will be written}
}
\value{
the information line for the "y name" field in the selected language
}
\description{
the description of the "y name" field
}

