% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-PlotGG-class.R
\name{addCustomPoints.PlotGG}
\alias{addCustomPoints.PlotGG}
\title{Add custom points for PlotGG}
\usage{
\method{addCustomPoints}{PlotGG}(plot_obj, custom_points, ...)
}
\arguments{
\item{plot_obj}{A plot-like object.}

\item{custom_points}{A list/data for points.}

\item{...}{Passed on to methods.}
}
\value{
A `PlotGG` object (modified).
}
\description{
Add custom points for PlotGG
}
