% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-formatGGplot.R
\name{addCustomPointsToGGplot}
\alias{addCustomPointsToGGplot}
\title{Add Custom Points To GGplot}
\usage{
addCustomPointsToGGplot(plot, custom_points)
}
\arguments{
\item{plot}{(ggplot)}

\item{custom_points}{(list) named list with point definitions, output of \code{customPointsServer}}
}
\description{
Add custom points to a ggplot.
}
