% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-formatGGplot.R
\name{addCustomPointsToGGplot}
\alias{addCustomPointsToGGplot}
\title{Add Custom Points To GGplot}
\usage{
addCustomPointsToGGplot(plot, custom_points)
}
\arguments{
\item{plot}{(ggplot)}

\item{data}{(data.frame) data to be plotted}

\item{pointStyle}{(list) named list with style definitions, or output of \code{plotPointsServer}}

\item{labelStyle}{(list) named list with style definitions, or output of \code{plotTextServer}}
}
\description{
Add custom points to a ggplot.
}
