% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotTitles.R
\name{defaultInitText}
\alias{defaultInitText}
\title{Default Init Text}
\usage{
defaultInitText(type = c("none", "ggplot", "base"))
}
\arguments{
\item{type}{(character) Type of the plot to add titles to, one of "none", "ggplot", "base".}
}
\description{
Initial list with default text elements
}
