% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-formatGGplot.R
\name{extractTitle}
\alias{extractTitle}
\title{Extract Title}
\usage{
extractTitle(titleList)
}
\arguments{
\item{titleList}{(list) named list with title definitions, output of \code{plotTitlesServer}}
}
\value{
(character) title
}
\description{
Extract title from list of title definitions
}
