% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/01-PlotGG-class.R
\name{formatLegend.PlotGG}
\alias{formatLegend.PlotGG}
\title{Format legend for PlotGG}
\usage{
\method{formatLegend}{PlotGG}(plot_obj, legend, scaleFUN = ggplot2::scale_color_manual, ...)
}
\arguments{
\item{plot_obj}{A plot-like object.}

\item{legend}{A legend configuration list.}

\item{scaleFUN}{A ggplot2 scale function (e.g., `ggplot2::scale_color_manual`).}

\item{...}{Passed on to methods.}
}
\value{
A `PlotGG` object (modified).
}
\description{
Format legend for PlotGG
}
