% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-formatGGplot.R
\name{formatLegendOfGGplot}
\alias{formatLegendOfGGplot}
\title{Legend Style Of GGplot}
\usage{
formatLegendOfGGplot(plot, legend, scaleFUN = ggplot2::scale_color_manual, ...)
}
\arguments{
\item{plot}{(ggplot) A ggplot object.}

\item{legend}{(list) named list with style definitions, or output of \code{plotLegendServer}}

\item{scaleFUN}{(function) function to set scale, e.g. \code{ggplot2::scale_color_manual}}

\item{...}{additional element specifications not part of base ggplot2. In general,
these should also be defined in the \verb{element tree} argument. \link[rlang:splice]{Splicing} a list is also supported.}
}
\value{
A ggplot object with updated legend.
}
\description{
Style of legend is defined with argument \code{legend}. Overwrites previous definitions of \code{theme(legend)}
}
