% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-formatGGplot.R
\name{formatPointsOfGGplot}
\alias{formatPointsOfGGplot}
\title{Point Style Of GGplot}
\usage{
formatPointsOfGGplot(plot, pointStyle)
}
\arguments{
\item{plot}{(ggplot)}

\item{pointStyle}{(list) named list with style definitions, or output of \code{plotPointsServer}}
}
\description{
Style of points is defined with \code{pointStyle}. Overwrites previous definitions of \code{geom_point}
}
