% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-formatGGplot.R
\name{formatRangesOfGGplot}
\alias{formatRangesOfGGplot}
\title{Axes Ranges Of GGplot}
\usage{
formatRangesOfGGplot(plot, ranges)
}
\arguments{
\item{plot}{(ggplot)}

\item{ranges}{(list) named list with range definitions, output of \code{plotRangesServer}}
}
\description{
Axes Ranges Of GGplot
}
