% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-formatGGPlot.R
\name{formatTitlesOfGGPlot}
\alias{formatTitlesOfGGPlot}
\title{Format Titles Of GGplot}
\usage{
formatTitlesOfGGPlot(plot, titles)
}
\arguments{
\item{plot}{(ggplot)}

\item{titles}{(list) named list with title definitions, output of \code{plotTitlesServer}}
}
\description{
Format Titles Of GGplot
}
