% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-formatGGplot.R
\name{formatTitlesOfGGplot}
\alias{formatTitlesOfGGplot}
\title{Format Titles Of GGplot}
\usage{
formatTitlesOfGGplot(plot, text)
}
\arguments{
\item{plot}{(ggplot) A ggplot object.}

\item{text}{(list) named list with title definitions, output of \code{plotTitlesServer}}
}
\value{
A ggplot object with titles formatted.
}
\description{
Format plot and axis titles and texts of axis labels (and legends) of a ggplot.
}
