% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-formatGGplot.R
\name{getElementText}
\alias{getElementText}
\title{Get Element Text}
\usage{
getElementText(
  textDef = list(fontFamily = "sans", hide = FALSE, size = 12L, fontType = "plain", color
    = "#000000", angle = 0, hjust = 0.5, vjust = 0.5)
)
}
\arguments{
\item{textDef}{(list) named list with specs for text formatting, see e.g.
\code{config()$defaultGGTitle}}
}
\description{
Get Element Text
}
