% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-headerButtons.R
\name{headerButtonsUI}
\alias{headerButtonsUI}
\title{Add header buttons}
\usage{
headerButtonsUI(
  id,
  help_link,
  further_help_link = NULL,
  loadShinyToolsCSS = TRUE
)
}
\arguments{
\item{id}{module id}

\item{help_link}{link that is opened when help button is clicked}

\item{further_help_link}{link that is opened when further help button is clicked; if set to NULL, no button is displayed}

\item{loadShinyToolsCSS}{logical should shinyToolsCSS be loaded}
}
\value{
tagList with divs
}
\description{
Add header buttons
}
