% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotExport.R
\name{plotExportServer}
\alias{plotExportServer}
\title{Server function for plot export}
\usage{
plotExportServer(
  id,
  plotFun,
  extraPlotFormatting = c("none", "ggplot"),
  filename = sprintf("\%s_plot", gsub("-", "", Sys.Date())),
  plotly = FALSE,
  plotWidth = reactive(1280),
  plotHeight = reactive(800)
)
}
\arguments{
\item{id}{namespace id}

\item{plotFun}{(reactive) a reactive function returning a plot for export}

\item{extraPlotFormatting}{(character) one of "none", "ggplot". Adds the option to format the
plot before export}

\item{filename}{(character) name of file without file extension}

\item{plotly}{(logical) set TRUE if plotFun returns a plotly output}

\item{plotWidth}{(reactive) default plot width}

\item{plotHeight}{(reactive) default plot height}
}
\description{
Backend for plot export module
}
