% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotExport.R
\name{plotExportServer}
\alias{plotExportServer}
\title{Server function for plot export}
\usage{
plotExportServer(
  id,
  plotFun,
  plotType = c("none", "ggplot", "ggplot_only_titles"),
  filename = sprintf("\%s_plot", gsub("-", "", Sys.Date())),
  plotly = FALSE,
  plotWidth = reactive(1280),
  plotHeight = reactive(800),
  initText = NULL,
  initRanges = NULL
)
}
\arguments{
\item{id}{namespace id}

\item{plotFun}{(reactive) a reactive function returning a plot for export}

\item{plotType}{(character) one of "none", "ggplot", "ggplot_only_titles". Adds the option to
format titles and ranges of a plot within the export UI (currently only for ggplots). For
\code{plotType == "ggplot_only_titles"} only titles can be adjusted. This prevents that custom
formatting of axis ranges might be overwritten by \code{formatScalesOfGGplot()}.}

\item{filename}{(character) name of file without file extension}

\item{plotly}{(logical) set TRUE if plotFun returns a plotly output}

\item{plotWidth}{(reactive) default plot width}

\item{plotHeight}{(reactive) default plot height}

\item{initText}{(list) optional, named list with title definitions, or output of \code{plotTitlesServer}}

\item{initRanges}{(list) optional, named list with range definitions, or output of \code{plotRangesServer}}
}
\description{
Backend for plot export module
}
