% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-plotLegend.R
\name{plotLegendUI}
\alias{plotLegendUI}
\alias{plotLegendServer}
\title{Plot Legend UI}
\usage{
plotLegendUI(
  id,
  title = NULL,
  titleTag = "h4",
  type = c("ggplot", "base", "none"),
  width = "100\%"
)

plotLegendServer(
  id,
  legend_title = reactive(NULL),
  legend_labels = reactive(NULL),
  plot_type = c("ggplot", "base", "none")
)
}
\arguments{
\item{id}{namespace id}

\item{title}{(character) title of the module}

\item{titleTag}{(character) HTML tag to put around the title, e.g. "h4" for \code{h4} from
\code{htmltools}}
}
\description{
Function to create the UI for the legend

Function to create the server for the legend
}
