% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/03-plotLegend.R
\name{plotLegendUI}
\alias{plotLegendUI}
\alias{plotLegendServer}
\title{Plot Legend UI}
\usage{
plotLegendUI(
  id,
  title = NULL,
  titleTag = "h4",
  type = c("ggplot", "base", "none"),
  width = "100\%"
)

plotLegendServer(
  id,
  legend_title = reactive(NULL),
  legend_labels = reactive(NULL),
  plot_type = c("ggplot", "base", "none")
)
}
\arguments{
\item{id}{namespace id}

\item{title}{(character) module title}

\item{titleTag}{(character) HTML tag to put around the title, e.g. "h4" for \code{h4} from
\code{htmltools}}

\item{type}{character. Type of plot. Default is "ggplot".}

\item{width}{character. Width of the input.}

\item{legend_title}{reactive. Default title of the legend.}

\item{legend_labels}{reactive. Default labels of the legend.}

\item{plot_type}{character. Type of plot. Default is "ggplot".}
}
\description{
Function to create the UI for the legend

Function to create the server for the legend
}
