% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotRanges.R
\name{plotRangesServer}
\alias{plotRangesServer}
\title{Server function for plot ranges}
\usage{
plotRangesServer(id, type = c("none", "ggplot", "base"), initRanges = NULL)
}
\arguments{
\item{id}{namespace id}

\item{type}{(character) Type of the plot to add ranges to, one of "ggplot", "base".}

\item{initRanges}{(list) optional, named list with range definitions, or output of \code{plotRangesServer}}
}
\description{
Backend for plot ranges module
}
