% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotRanges.R
\name{plotRangesServer}
\alias{plotRangesServer}
\title{Server function for plot ranges}
\usage{
plotRangesServer(id, type = c("none", "ggplot", "base"), ranges = NULL)
}
\arguments{
\item{id}{namespace id}

\item{type}{(character) Type of the plot to add ranges to, one of "ggplot", "base".}

\item{ranges}{(reactiveValues) initial ranges to be used when loading the plot}
}
\description{
Backend for plot ranges module
}
