% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotRanges.R
\name{plotRangesUI}
\alias{plotRangesUI}
\title{Plot Ranges UI}
\usage{
plotRangesUI(id, title = "Ranges", titleTag = "h4", initRanges = deprecated())
}
\arguments{
\item{id}{module id}

\item{title}{(character) module title}

\item{titleTag}{(character) HTML tag to put around the title, e.g. "h4" for \code{h4} from
\code{htmltools}}

\item{initRanges}{(Deprecated) This parameter is no longer needed and has no effect.}
}
\value{
tagList
}
\description{
Plot Ranges UI
}
