% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotTitles.R
\name{plotTitlesServer}
\alias{plotTitlesServer}
\title{Server function for plot titles}
\usage{
plotTitlesServer(id, type = c("none", "ggplot", "base"), titles = NULL)
}
\arguments{
\item{id}{namespace id}

\item{type}{(character) Type of the plot to add titles to, one of "ggplot", "base".}

\item{titles}{(reactiveValues) initial titles to be used when loading the plot}
}
\description{
Backend for plot titles module
}
