% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotTitles.R
\name{plotTitlesUI}
\alias{plotTitlesUI}
\title{Plot Titles UI}
\usage{
plotTitlesUI(id, type = c("ggplot", "base"))
}
\arguments{
\item{id}{module id}

\item{type}{(character) Type of the plot to add titles to, one of "ggplot", "base".}
}
\value{
tagList
}
\description{
Plot Titles UI
}
