% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-plotTitles.R
\name{plotTitlesUI}
\alias{plotTitlesUI}
\title{Plot Titles UI}
\usage{
plotTitlesUI(id, extraPlotFormatting = c("none", "ggplot"))
}
\arguments{
\item{id}{module id}

\item{extraPlotFormatting}{(character) one of "none", "ggplot". Adds the option to format the
plot before export}
}
\value{
tagList
}
\description{
Plot Titles UI
}
