% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/02-vectorInput.R
\name{vectorInputUI}
\alias{vectorInputUI}
\title{Vector Input UI}
\usage{
vectorInputUI(id, newValueRange, title = NULL, titleTag = "strong")
}
\arguments{
\item{id}{module id}

\item{newValueRange}{(numeric) Range of values that can be entered in the numeric input field}

\item{title}{(character) module title}

\item{titleTag}{(character) HTML tag to put around the title, e.g. "h4" for \code{h4} from
\code{htmltools}}
}
\value{
tagList
}
\description{
Vector Input UI
}
